pyproject.toml
[tool.poetry]
name = "tradingview-notify"
version = "0.1.0"
description = ""
authors = ["andy010629 <andy010629@gmail.com>"]
readme = "README.md"
packages = [{include = "tradingview_notify"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
dotenv-python = "^0.0.1"
fastapi = "^0.100.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
